# docs/swagger.yaml
basePath: /
definitions:
  models.LoginForm:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.LoraTextRequest:
    properties:
      prompt:
        type: string
      task_id:
        type: integer
    type: object
  models.RegisterForm:
    properties:
      confirm_password:
        type: string
      email:
        description: 邮箱
        type: string
      gender:
        description: 性别 0:未知 1:男 2:女
        enum:
        - 0
        - 1
        - 2
        type: integer
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    required:
    - confirm_password
    - email
    - password
    - username
    type: object
  models.T2IRequest:
    properties:
      task_id:
        type: integer
    type: object
  models.V2TRequest:
    properties:
      video_url:
        type: string
    type: object
  task.I2VTask:
    properties:
      created_at:
        type: integer
      image_url:
        type: string
      index:
        type: integer
      priority:
        type: integer
      prompt:
        type: string
      task_id:
        type: integer
      user_id:
        type: integer
    type: object
host: 198.168.1.50:8080
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: 视频处理相关 API 接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: V2V API
  version: "1.0"
paths:
  /I2V/{task_id}:
    get:
      consumes:
      - application/json
      description: 通过任务 ID 获取 I2V 任务（输入任务I2V的ID可以查询到任务完成情况）
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"video_url": "..."}'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取 I2V 任务结果
      tags:
      - I2V
  /api/v1/FFmpeg/{task_id}:
    get:
      consumes:
      - application/json
      description: 处理 FFmpeg 相关的视频处理任务 （在一切视频都生成后输入任务I2V的ID进行拼接）
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "..."}'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: FFmpeg 处理器
      tags:
      - FFmpeg
  /api/v1/I2V:
    post:
      consumes:
      - application/json
      description: 接收参考图片和文本提示词，创建一个新的 I2V 任务并返回任务 ID （输入从T2I获得的任务I2V的ID）
      parameters:
      - description: I2V 任务请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/task.I2VTask'
      produces:
      - application/json
      responses:
        "202":
          description: '{"task_id": 123456, "status": "task submitted"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 提交图片生成视频任务
      tags:
      - I2V
  /api/v1/I2VCallback/{task_id}:
    post:
      consumes:
      - application/json
      description: 处理来自视频生成服务的任务完成回调
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      - description: Callback Data
        in: body
        name: data
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"status": "success"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: I2V 任务回调处理
      tags:
      - I2V
  /api/v1/T2I:
    post:
      consumes:
      - application/json
      description: 接收任务ID，创建一个新的 T2I 任务并返回任务 ID（输入从V2T得到的任务ID）
      parameters:
      - description: T2I 任务请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.T2IRequest'
      produces:
      - application/json
      responses:
        "202":
          description: '{"task_id": 123456, "status": "task submitted"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 提交文本生成图片任务
      tags:
      - T2I
  /api/v1/V2T:
    post:
      consumes:
      - application/json
      description: 接收视频URL，创建一个新的 V2T 任务并返回任务 ID （目前没有登陆注册，所以就只要求输入视频链接）
      parameters:
      - description: V2T 任务请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.V2TRequest'
      produces:
      - application/json
      responses:
        "202":
          description: '{"task_id": "123456", "status": "submitted"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 提交视频转文字任务
      tags:
      - V2T
  /api/v1/V2T/{task_id}:
    get:
      consumes:
      - application/json
      description: 通过任务 ID 获取 V2T 任务的当前状态和结果 （输入任务ID即可）
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"task_id": 123456, "status": "completed", "result": "..."}'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: task not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取 V2T 任务结果
      tags:
      - V2T
  /api/v1/V2T/LoraText:
    post:
      consumes:
      - application/json
      description: 为指定任务更新 Lora 相关的文本提示词 （同时输入任务ID与更新后的提示词即可）
      parameters:
      - description: Lora 文本更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoraTextRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"task_id": 123456, "status": "task updated"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新任务 Lora 文本
      tags:
      - V2T
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录，返回 access_token 和 refresh_token
      parameters:
      - description: 登录表单（username 和 password）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginForm'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回用户ID、用户名和两个token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 用户不存在或密码错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 用户登录
      tags:
      - Auth
  /api/v1/refresh_token:
    post:
      consumes:
      - application/json
      description: 使用 refresh_token 刷新 access_token，需在 Authorization 请求头中提供 Bearer
        token
      parameters:
      - description: 刷新令牌（也可在 Authorization header 中提供）
        in: query
        name: refresh_token
        required: true
        type: string
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功，返回新的 access_token 和 refresh_token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Token 格式错误或缺失
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Token 无效或过期
          schema:
            additionalProperties: true
            type: object
      summary: 刷新访问令牌
      tags:
      - Auth
  /api/v1/signup:
    post:
      consumes:
      - application/json
      description: 创建新用户账号，返回标准响应体
      parameters:
      - description: 注册表单（username 和 password）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterForm'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误或用户已存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 用户注册
      tags:
      - Auth
  /events:
    get:
      consumes:
      - application/json
      description: 建立 SSE 长连接以接收服务端推送的事件。需要通过查询参数 `userid` 指定订阅的主题/用户 ID，例如 `/events?userid=12345`。会在V2T，T2I任务结束后推送消息。
      parameters:
      - description: User ID / topic to subscribe
        in: query
        name: userid
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: event stream
          schema:
            type: string
        "400":
          description: missing topic
          schema:
            type: string
        "500":
          description: server error
          schema:
            type: string
      summary: 订阅服务器事件流（SSE）
      tags:
      - SSE
schemes:
- http
- https
swagger: "2.0"